Combine all cooldown commands
This effectively reverts 0fcb77a, where cooldown commands were generated
for EACH listener where specified. This doesn't scale and pollutes the
command menu. This commit fixes this by having just two commands,
/set-listener-cooldown and /override-listener-cooldown, that accept a
listener ID and then perform the cooldown updating based on that.

Also, 0fcb77a actually missed a bug where using CooldownManager#set
discards the onCooldown property, which I had forgotten existed. The
result was that the Klee dab feature stopped reacting with neko L after
using an cooldown update command. This commit fixes this by introducing
a new CooldownManager#update method, which serves as the "Update"
counterpart to set()'s "Create" in the CooldownSpec CRUD.
