+++ src/controllers/special/luke.controller.ts
@@ -73,6 +73,6 @@ const spec = new Controller({
   name: "luke",
   commands: [setMeowChance],
   listeners: [dadJoker, randomMeower],
-}).withCooldownCommands();
+});
 
 export default spec;
