+++ src/middleware/cooldown.middleware.ts
@@ -1,17 +1,11 @@
 import {
   Awaitable,
-  CommandInteractionOptionResolver,
-  GuildMember,
-  Message,
-  Role,
-  SlashCommandBuilder,
+  Message
 } from "discord.js";
 import lodash from "lodash";
 
 import getLogger from "../logger";
-import { Command } from "../types/command.types";
 import { addDateSeconds, formatHoursMinsSeconds } from "../utils/dates.utils";
-import { getAllMembers } from "../utils/iteration.utils";
 import {
   joinUserMentions,
   toBulletedList,
@@ -19,7 +13,6 @@ import {
   toTimestampMention,
   toUserMention,
 } from "../utils/markdown.utils";
-import { RoleLevel, checkPrivilege } from "./privilege.middleware";
 
 const log = getLogger(__filename);
 
@@ -102,6 +95,18 @@ export class CooldownManager {
       this.setBypass(true, memberId);
     }
   };
+
+  public update = (spec: Partial<CooldownSpec>): void => {
+    if (!this.spec) {
+      const message = "attempted to update a spec that hasn't been set yet";
+      log.error(`${message}.`);
+      throw new Error(message);
+    }
+
+    // Overwrite only the keys that are provided in the new spec.
+    const updatedCompleteSpec = { ...this.spec, ...spec } as CooldownSpec;
+    this.set(updatedCompleteSpec);
+  }
 
   public setDuration(seconds: number): void;
   public setDuration(seconds: number, userId: string): void;
@@ -325,167 +330,5 @@ export class CooldownManager {
     }
 
     return null;
-  };
-
-  public getCooldownSetterCommand = (listenerId: string): Command => {
-    const command = new Command(new SlashCommandBuilder()
-      .setName(`set-${listenerId}-cooldown`)
-      .setDescription(`Set the cooldown spec for the ${listenerId} listener.`)
-      .addStringOption(input => input
-        .setName("type")
-        .setDescription("Cooldown type.")
-        .setRequired(true)
-        .addChoices(
-          { name: "Global", value: "global" },
-          { name: "Per-user", value: "user" },
-          { name: "Disabled", value: "disabled" },
-        )
-      )
-      .addNumberOption(input => input
-        .setName("seconds")
-        .setDescription("Default duration of cooldown (in seconds).")
-        .setMinValue(0)
-      ),
-      { broadcastOption: true },
-    );
-
-    command.check(checkPrivilege(RoleLevel.DEV));
-    command.execute(async (interaction) => {
-      const options = interaction.options as CommandInteractionOptionResolver;
-      const broadcast = options.getBoolean("broadcast") ?? false;
-      const type = options.getString("type", true) as CooldownSpec["type"];
-
-      if (type === "disabled") {
-        this.set({ type: "disabled" });
-        await interaction.reply({
-          content: `Disabled cooldown for **${listenerId}**!`,
-          ephemeral: !broadcast,
-        });
-        return;
-      }
-
-      const seconds = options.getNumber("seconds");
-      if (seconds === null) {
-        await interaction.reply({
-          content: "Specify a value for the default cooldown duration!",
-          ephemeral: true,
-        });
-        return;
-      }
-
-      if (type === "global") {
-        this.set({ type: "global", seconds });
-      } else if (type === "user") {
-        this.set({ type: "user", defaultSeconds: seconds });
-      }
-
-      const response = `Updated **${listenerId}** cooldown spec:\n` +
-        toBulletedList([
-          `Type: \`${type}\``,
-          `Default duration: ${formatHoursMinsSeconds(seconds)}`,
-        ]);
-      await interaction.reply({ content: response, ephemeral: !broadcast });
-    });
-
-    return command;
-  };
-
-  public getCooldownOverriderCommand = (listenerId: string): Command => {
-    const command = new Command(new SlashCommandBuilder()
-      .setName(`override-${listenerId}-cooldown`)
-      .setDescription(`Set cooldown duration overrides for ${listenerId}.`)
-      .addMentionableOption(input => input
-        .setName("mentionable")
-        // TODO: supporting roles mean that the underlying list of members with
-        // overrides can easily explode in size, likely making the current
-        // implementation of /cooldowns (which returns an unpaginated text
-        // response) error out from exceeded message length limit.
-        .setDescription("Member or role to set overrides for.")
-        .setRequired(true)
-      )
-      .addNumberOption(input => input
-        .setName("duration")
-        .setDescription(
-          "User-specific cooldown duration override " +
-          "(in seconds) (USER type cooldown only)."
-        )
-        .setMinValue(0)
-      )
-      .addBooleanOption(input => input
-        .setName("bypass")
-        .setDescription("Allow this user to bypass cooldown duration.")
-      ),
-      { broadcastOption: true },
-    );
-
-    command.check(checkPrivilege(RoleLevel.DEV));
-    command.execute(async (interaction) => {
-      const options = interaction.options as CommandInteractionOptionResolver;
-      const broadcast = options.getBoolean("broadcast") ?? false;
-
-      if (this.spec.type === "disabled") {
-        await interaction.reply({
-          content: `Cooldown for **${listenerId}** is currently disabled!`,
-          ephemeral: true,
-        });
-        return;
-      }
-
-      const mentionable
-        = options.getMentionable("mentionable", true) as GuildMember | Role;
-      const members = getAllMembers(mentionable);
-
-      const duration = options.getNumber("duration");
-      const bypass = options.getBoolean("bypass");
-
-      if (duration !== null && bypass !== null) {
-        await interaction.reply({
-          content: "Provide a duration or bypass value but not both!",
-          ephemeral: true,
-        });
-        return;
-      }
-
-      if (duration !== null) {
-        if (this.spec.type !== "user") {
-          await interaction.reply({
-            content: "`duration` is only compatible with `user` cooldown type.",
-            ephemeral: true,
-          });
-          return;
-        }
-        for (const member of members)
-          this.setDuration(duration, member.id);
-        await interaction.reply({
-          content:
-            `Set **${listenerId}** cooldown duration override for ` +
-            `${mentionable}: ${formatHoursMinsSeconds(duration)}.`,
-          ephemeral: !broadcast,
-          // TODO: Maybe make a replySilently overload for interactions.
-          allowedMentions: { parse: [] },
-        });
-        return;
-      }
-
-      if (bypass !== null) {
-        for (const member of members)
-          this.setBypass(bypass, member.id);
-        await interaction.reply({
-          content:
-            `${bypass ? "Enabled" : "Disabled"} **${listenerId}** bypass ` +
-            `for ${mentionable}.`,
-          ephemeral: !broadcast,
-          allowedMentions: { parse: [] },
-        });
-        return;
-      }
-
-      await interaction.reply({
-        content: "Missing argument(s)!",
-        ephemeral: true,
-      });
-    });
-
-    return command;
   };
 }
